requirements.txt
dataclasses
transformers
redis
sentence-transformers
torch
ollama
typer
pyyaml